refactor(login): move authentication to Login component
	* Move authentication from url paramter to Login component
	instead of using own route and component.
	* Move up debt collection view over other registration views
