+++ src/modules/login/views/index.ts
@@ -1,4 +1,3 @@
 export * from './Login'
 export * from './ForgotPassword'
 export * from './ChangePasswordWithToken'
-export * from './Authenticate'
