+++ src/routes/PublicRoutes.tsx
@@ -4,7 +4,6 @@ import {
   ChangePasswordWithToken,
   ForgotPassword,
   Login,
-  Authenticate,
 } from 'modules/login/views'
 import { ErrorBoundary } from 'react-error-boundary'
 import { Route, Routes } from 'react-router-dom'
@@ -18,20 +17,6 @@ const PublicRoutes: React.FC = () => {
       </Route>
       <Route path="forgot-password" element={<ForgotPassword />} />
       <Route path="reset-password" element={<ChangePasswordWithToken />} />
-
-      {/* authenticate endpoint is used to load a jwt token sent from backend emails to log in */}
-
-      <Route
-        path="authenticate"
-        element={
-          <ErrorBoundary
-            FallbackComponent={() => <div>Oida, her gikk noe galt</div>}
-          >
-            <Authenticate />
-          </ErrorBoundary>
-        }
-      />
-
       <Route path="login" element={<Login />} />
       <Route path="*" element={<Login />} />
     </Routes>
